ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pdfplumber==0.11.0
python-docx==1.1.0
spacy==3.7.4
typer==0.9.4
sentence-transformers==2.7.0
python-multipart==0.0.9
